fix(admin-api): delete references to krt v2 in validator
