+++ engine/admin-api/domain/usecase/krt/validator/fields_validator_test.go
@@ -17,43 +17,43 @@ func TestYamlFieldsValidator_Run(t *testing.T) {
 	}{
 		{
 			name:        "KRT YAML values successfully validated",
-			krtYaml:     NewKrtBuilder().V2().Build(),
+			krtYaml:     NewKrtBuilder().Build(),
 			wantError:   false,
 			errorString: "",
 		},
 		{
 			name:        "fails if version name has an invalid length",
-			krtYaml:     NewKrtBuilder().V2().WithVersion("this-version-name-length-is-higher-than-the-maximum").Build(),
+			krtYaml:     NewKrtBuilder().WithVersion("this-version-name-length-is-higher-than-the-maximum").Build(),
 			wantError:   true,
 			errorString: "invalid length \"this-version-name-length-is-higher-than-the-maximum\" at \"Version\" must be lower than 20",
 		},
 		{
 			name:        "fails if krt hasn't required field description",
-			krtYaml:     NewKrtBuilder().V2().WithDescription("").Build(),
+			krtYaml:     NewKrtBuilder().WithDescription("").Build(),
 			wantError:   true,
 			errorString: "the field \"Description\" is required",
 		},
 		{
 			name:        "fails if proto doesn't end with .proto",
-			krtYaml:     NewKrtBuilder().V2().WithEntrypointProto("invalid").Build(),
+			krtYaml:     NewKrtBuilder().WithEntrypointProto("invalid").Build(),
 			wantError:   true,
 			errorString: "invalid value \"invalid\" at \"Entrypoint.Proto\" must end with .proto",
 		},
 		{
 			name:        "fails if required proto field is not defined",
-			krtYaml:     NewKrtBuilder().V2().WithEntrypointProto("").Build(),
+			krtYaml:     NewKrtBuilder().WithEntrypointProto("").Build(),
 			wantError:   true,
 			errorString: "the field \"Entrypoint.Proto\" is required",
 		},
 		{
 			name:        "fails if required image field is not defined",
-			krtYaml:     NewKrtBuilder().V2().WithEntrypointImage("").Build(),
+			krtYaml:     NewKrtBuilder().WithEntrypointImage("").Build(),
 			wantError:   true,
 			errorString: "the field \"Entrypoint.Image\" is required",
 		},
 		{
 			name:        "fails if and invalid env var is defined in config",
-			krtYaml:     NewKrtBuilder().V2().WithConfigVars([]string{"invalid-var"}).Build(),
+			krtYaml:     NewKrtBuilder().WithConfigVars([]string{"invalid-var"}).Build(),
 			wantError:   true,
 			errorString: "invalid value \"invalid-var\" at env var \"Config.Variables[0]\" must contain only capital letters, numbers, and underscores",
 		},
