+++ engine/admin-api/domain/usecase/krt/validator/validator_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/konstellation-io/kre/engine/admin-api/mocks"
 )
 
-func TestValidatorV2_Run(t *testing.T) {
+func TestKrtValidator_Run(t *testing.T) {
 	tests := []struct {
 		name        string
 		krtYaml     *krt.Krt
@@ -21,13 +21,13 @@ func TestValidatorV2_Run(t *testing.T) {
 	}{
 		{
 			name:        "valid workflow validation",
-			krtYaml:     NewKrtBuilder().V2().Build(),
+			krtYaml:     NewKrtBuilder().Build(),
 			wantError:   false,
 			errorString: "",
 		},
 		{
 			name: "fails if some node have no subscriptions",
-			krtYaml: NewKrtBuilder().V2().WithWorkflowsNodes([]krt.Node{
+			krtYaml: NewKrtBuilder().WithWorkflowsNodes([]krt.Node{
 				{
 					Name:  "test-node",
 					Image: "test-image",
@@ -39,7 +39,7 @@ func TestValidatorV2_Run(t *testing.T) {
 		},
 		{
 			name: "fails if some workflow hasn't exitpoint defined",
-			krtYaml: NewKrtBuilder().V2().WithWorkflows([]krt.Workflow{{
+			krtYaml: NewKrtBuilder().WithWorkflows([]krt.Workflow{{
 				Name:       "test-workflow",
 				Entrypoint: "test-entrypoint",
 				Nodes: []krt.Node{{
@@ -55,14 +55,13 @@ func TestValidatorV2_Run(t *testing.T) {
 		},
 		{
 			name:        "fails if exitpoint node doesn't exist in nodes list",
-			krtYaml:     NewKrtBuilder().V2().WithWorkflowsExitpoint("inexistent-node").Build(),
+			krtYaml:     NewKrtBuilder().WithWorkflowsExitpoint("inexistent-node").Build(),
 			wantError:   true,
 			errorString: "exitpoint node \"inexistent-node\" not found in workflow \"valid-workflow\" nodes",
 		},
 		{
 			name: "fails if node name is not unique in workflow",
 			krtYaml: NewKrtBuilder().
-				V2().
 				WithWorkflowsNodes([]krt.Node{
 					{
 						Name:          "test-node",
@@ -89,7 +88,7 @@ func TestValidatorV2_Run(t *testing.T) {
 
 	mocks.AddLoggerExpects(logger)
 
-	validator := validator.NewValidator(logger, fieldsValidator)
+	validator := validator.NewKrtValidator(logger, fieldsValidator)
 
 	for _, tc := range tests {
 		fieldsValidator.EXPECT().Run(tc.krtYaml).Return(nil)
